:recycle: refactor JSON serialization functions to private module
