+++ discord/gateway.py
@@ -42,6 +42,7 @@ from . import utils
 from .activity import BaseActivity
 from .enums import SpeakingState
 from .errors import ConnectionClosed, InvalidArgument
+from .utils.private import _from_json, _to_json
 
 _log = logging.getLogger(__name__)
 
@@ -450,7 +451,7 @@ class DiscordWebSocket:
             self._buffer = bytearray()
 
         self.log_receive(msg)
-        msg = utils._from_json(msg)
+        msg = _from_json(msg)
 
         _log.debug("For Shard ID %s: WebSocket Event: %s", self.shard_id, msg)
         event = msg.get("t")
@@ -637,7 +638,7 @@ class DiscordWebSocket:
 
     async def send_as_json(self, data):
         try:
-            await self.send(utils._to_json(data))
+            await self.send(_to_json(data))
         except RuntimeError as exc:
             if not self._can_handle_close():
                 raise ConnectionClosed(self.socket, shard_id=self.shard_id) from exc
@@ -645,7 +646,7 @@ class DiscordWebSocket:
     async def send_heartbeat(self, data):
         # This bypasses the rate limit handling code since it has a higher priority
         try:
-            await self.socket.send_str(utils._to_json(data))
+            await self.socket.send_str(_to_json(data))
         except RuntimeError as exc:
             if not self._can_handle_close():
                 raise ConnectionClosed(self.socket, shard_id=self.shard_id) from exc
@@ -671,7 +672,7 @@ class DiscordWebSocket:
             },
         }
 
-        sent = utils._to_json(payload)
+        sent = _to_json(payload)
         _log.debug('Sending "%s" to change status', sent)
         await self.send(sent)
 
@@ -774,7 +775,7 @@ class DiscordVoiceWebSocket:
 
     async def send_as_json(self, data):
         _log.debug("Sending voice websocket frame: %s.", data)
-        await self.ws.send_str(utils._to_json(data))
+        await self.ws.send_str(_to_json(data))
 
     send_heartbeat = send_as_json
 
@@ -931,7 +932,7 @@ class DiscordVoiceWebSocket:
         # This exception is handled up the chain
         msg = await asyncio.wait_for(self.ws.receive(), timeout=30.0)
         if msg.type is aiohttp.WSMsgType.TEXT:
-            await self.received_message(utils._from_json(msg.data))
+            await self.received_message(_from_json(msg.data))
         elif msg.type is aiohttp.WSMsgType.ERROR:
             _log.debug("Received %s", msg)
             raise ConnectionClosed(self.ws, shard_id=None) from msg.data
