+++ discord/utils/private.py
@@ -530,3 +530,20 @@ def cached_slot_property(
         return CachedSlotProperty(name, func)
 
     return decorator
+
+
+try:
+    import msgspec
+
+    def _to_json(obj: Any) -> str:  # type: ignore
+        return msgspec.json.encode(obj).decode("utf-8")
+
+    _from_json = msgspec.json.decode  # type: ignore
+
+except ModuleNotFoundError:
+    import json
+
+    def _to_json(obj: Any) -> str:
+        return json.dumps(obj, separators=(",", ":"), ensure_ascii=True)
+
+    _from_json = json.loads
